Update Report.tex
Made significant changes not just added hw 3, but also reorganized the structure a bit based on the feedback I recieved from homework 1. Also forgot to add some bits from homework 2 because I uploaded a previous version so brought that up to date. Also included some of my notes and thoughts from lectures on there as well and adding summaries about the homework assignments and about what they are about. Also went into a more in depth explanation of the MIU puzzle as part of the feedback on the assingment as well. Overall a much cleaner version. Basically the majority was reworked from scratch